Created and converted mixins
