+++ package.json
@@ -55,7 +55,7 @@
     "nyc": "^14.1.1",
     "sinon-chai": "^3.3.0",
     "ts-sinon": "^1.0.17",
-    "tslint": "^5.17.0",
+    "tslint": "^5.20.0",
     "tslint-config-airbnb": "^5.11.1",
     "typescript": "^3.5.1"
   },
