+++ src/client/mixins/index.ts
@@ -1,0 +1,3 @@
+export * from './basic-interaction';
+export * from './dom';
+export * from './response';
